% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{append_clc_classes_buffer}
\alias{append_clc_classes_buffer}
\title{Appends clc reclassified area percentage columns to a sf data frame}
\usage{
append_clc_classes_buffer(corine.wal.simple.sf, radius.num, locations.sf)
}
\arguments{
\item{corine.wal.simple.sf}{A simple feature of land covers in Wallonia}

\item{radius.num}{A numeric corresponding to the radius of the buffer that you want}

\item{locations.sf}{A simple feature which has coordinates of points}
}
\value{
a sf data frame presenting percentage of cover of each land cover for each location
}
\description{
Inspired from https://gis.stackexchange.com/questions/229453/create-a-circle-of-defined-radius-around-a-point-and-then-find-the-overlapping-a and
https://stackoverflow.com/questions/46704878/circle-around-a-geographic-point-with-st-buffer
}
