% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{build.clc_wallonia.polygons_sf.fun}
\alias{build.clc_wallonia.polygons_sf.fun}
\title{Reclassify the Corine Land Cover classes in custom & more general classes}
\usage{
build.clc_wallonia.polygons_sf.fun()
}
\value{
a sf dataframe containing the reclassified clc polygons
}
\description{
Inspired from https://stackoverflow.com/questions/13982773/crop-for-spatialpolygonsdataframe
}
\author{
Loïc Davadan - ldavadan.github.io
}
