% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation_and_manipulation.R
\name{compute_lm}
\alias{compute_lm}
\title{Compute a simple linear model and return its statistics in a dataframe}
\usage{
compute_lm(records.wide.df, output)
}
\arguments{
\item{records.wide.df}{A wide dataframe containing the records you wan to submit to lm}

\item{output}{a character specifying the type of output you want ("lm.sm" or "lm")}
}
\value{
a dataframe, either from lm or lm.sum
}
\description{
Compute a simple linear model and return its statistics in a dataframe
}
\author{
Thomas Goossens - pokyah.github.io
}
