% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{convert_stations_clc_buffer}
\alias{convert_stations_clc_buffer}
\title{It reshapes data::todo::}
\usage{
convert_stations_clc_buffer(class.buffers.sf = NULL)
}
\arguments{
\item{class.buffers.sf}{A simple feature corresponding to buffers associated to their land covers with proportions}
}
\value{
a data frame with percentage of each land covers for each station
}
\description{
It reshapes data::todo::
}
