% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agromet_API.R
\name{get_from_agromet_API.fun}
\alias{get_from_agromet_API.fun}
\title{Retrieve data from the [AGROMET API](https://app.pameseb.be/fr/pages/api_call_test/).}
\usage{
get_from_agromet_API.fun(user_token.chr = NULL, table_name.chr = NULL,
  sensors.chr = NULL, stations_ids.chr, dfrom.chr = NULL, dto.chr = NULL,
  month_day.chr = NULL, api_v.chr = NULL, test.bool = FALSE)
}
\arguments{
\item{user_token.chr}{A character specifying your user token}

\item{table_name.chr}{A character specifying the database table you want to query}

\item{sensors.chr}{A character vector containing the names of the sensors you want to query}

\item{stations_ids.chr}{A character vector containing the ids of the stations you want to query}

\item{dfrom.chr}{A character specifying the data from which you want to query data (YYYY-MM-DD)}

\item{dto.chr}{A character specifying the data to which you want to query data (YYYY-MM-DD)}

\item{month_day.chr}{A character specifying the month and day from which you want TMY data (MM-DD)}

\item{api_v.chr}{A character specifying the version of the API you want to use ("v1" or "v2")}

\item{test.bool}{A boolean, TRUE if you want to query the test server}

\item{user_token.chr}{A character specifying your user token}
}
\value{
A list containing the metadata in the first element and the queried data in the second element
}
\description{
Retrieve data from the [AGROMET API](https://app.pameseb.be/fr/pages/api_call_test/).
}
