% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{get_installed_packages.fun}
\alias{get_installed_packages.fun}
\title{List all the packages installed by user.}
\usage{
get_installed_packages.fun()
}
\value{
a dataframe containing all the installed packages and their version
}
\description{
Inspired from https://stackoverflow.com/questions/38481980/get-the-list-of-installed-packages-by-user-in-r#40120266
}
\author{
Thomas Goossens - pokyah.github.io
}
