% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation_and_manipulation.R
\name{h.check_NA}
\alias{h.check_NA}
\title{Check for NA values in a dataframe and log rows with NA values in a dataframe}
\usage{
h.check_NA(records.df)
}
\arguments{
\item{records.df}{a dataframe containing the records you want to check for NA's}
}
\value{
a dataframe containing al lthe records with at least one NA value
}
\description{
Check for NA values in a dataframe and log rows with NA values in a dataframe
}
\author{
Thomas Goossens - pokyah.github.io
}
