% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation_and_manipulation.R
\name{h.compute_stats}
\alias{h.compute_stats}
\title{Compute the summary stats for the desired sensor}
\usage{
h.compute_stats(input_records.df, sensor_name.chr)
}
\arguments{
\item{input_records.df}{A dataframe containing the records on which you want to compute the summary stats for a sensor}

\item{sensor_name.chr}{a character specifying the name of the sensor on which you want to compute the summary stats}
}
\value{
a dataframe containing the summary stats of the specified sensor
}
\description{
Compute the summary stats for the desired sensor
}
\author{
Thomas Goossens - pokyah.github.io
}
