% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation_and_manipulation.R
\name{h.filter_records}
\alias{h.filter_records}
\title{Filter a records dataset according to the passed pre-defined filtering parameter}
\usage{
h.filter_records(records.df, sensor.chr, dateRange.chr, filter.chr)
}
\arguments{
\item{records.df}{A dataframe containing the records you want filter}

\item{sensor.chr}{a character specifying the name of the sensor for which you want to apply a pre-defined filter}

\item{dateRange.chr}{a character vector specifying the dateRange on which you want to filter the dataset}

\item{filter.chr}{a character specifying the pre-defined filter you want to apply to your dataframe}
}
\value{
a filtered dataframe
}
\description{
Filter a records dataset according to the passed pre-defined filtering parameter
}
\author{
Thomas Goossens - pokyah.github.io
}
