% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation_and_manipulation.R
\name{h.get_date}
\alias{h.get_date}
\title{Get the widest shared date range of 2 AWS}
\usage{
h.get_date(records.df, s1, s2, type)
}
\arguments{
\item{records.df}{a dataframe containing the records you want filter}

\item{s1}{a character specifying the id of the first station}

\item{s2}{a character specifying the id of the second station}

\item{type}{a character specifying if max date or min date ("min" ou "max)}
}
\value{
a character containing the id of the station
}
\description{
Get the widest shared date range of 2 AWS
}
\author{
Thomas Goossens - pokyah.github.io
}
