% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation_and_manipulation.R
\name{h.get_id}
\alias{h.get_id}
\title{Get sid of sation based on its name_id}
\usage{
h.get_id(station_name_id.chr)
}
\arguments{
\item{station_name_id.chr}{a character specifying the name_sid of the station yo uwant to obtain the id from}
}
\value{
a character containing the id of the station
}
\description{
Get sid of sation based on its name_id
}
\author{
Thomas Goossens - pokyah.github.io
}
