% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_presentation.R
\name{h.ggplot_colours}
\alias{h.ggplot_colours}
\title{Create a color palette based on the default ggplot colours}
\usage{
h.ggplot_colours(n = 6, h = c(0, 360) + 15)
}
\value{
a colour palette
}
\description{
Inspired from https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}
\author{
Thomas Goossens - pokyah.github.io
}
