% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation_and_manipulation.R
\name{h.is_it_day}
\alias{h.is_it_day}
\title{Append a boolean column for is it day or not on an input dataframe containing a mtime column}
\usage{
h.is_it_day(records.df)
}
\arguments{
\item{records.df}{a dataframe containing a mtime column}
}
\value{
The same dataframe with a new boolean is it day column
}
\description{
Append a boolean column for is it day or not on an input dataframe containing a mtime column
}
\author{
Thomas Goossens - pokyah.github.io
}
