% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation_and_manipulation.R
\name{h.make_wide}
\alias{h.make_wide}
\title{Transform a records dataframe from long to wide format}
\usage{
h.make_wide(input_records.df, sensor_name.chr)
}
\arguments{
\item{input_records.df}{A dataframe containing the records you want to reshape in wide format}

\item{sensor_name.chr}{a character specifying the name of the sensor you want to reshape in wide format}
}
\value{
a wide dataframe containing the records for the specified sensor
}
\description{
Transform a records dataframe from long to wide format
}
\author{
Thomas Goossens - pokyah.github.io
}
