% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation_and_manipulation.R
\name{h.rad_top_atm}
\alias{h.rad_top_atm}
\title{Compute the value of Clearness index on the basis of mtime, lon and lat and append it as a new column}
\usage{
h.rad_top_atm(records.df)
}
\arguments{
\item{records.df}{A dataframe containing the records from the stations}
}
\value{
The dataframe with a newly computed CI column
}
\description{
Compute the value of Clearness index on the basis of mtime, lon and lat and append it as a new column
}
\author{
Thomas Goossens - pokyah.github.io & Michel Journée (RMI)
}
