% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation_and_manipulation.R
\name{lm_output_to_df.fun}
\alias{lm_output_to_df.fun}
\title{Transform a linear model list output to a dataframe. Useful for markdown export with [kable](https://www.rdocumentation.org/packages/knitr/versions/1.20/topics/kable)}
\usage{
lm_output_to_df.fun(lm.l)
}
\arguments{
\item{lm.l}{an object produced by lm()}
}
\value{
a dataframe presenting the model outputs
}
\description{
Inspired from http://r.789695.n4.nabble.com/Export-summary-from-regression-output-td4647109.html
}
\author{
Thomas Goossens - pokyah.github.io
}
