% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agromet_API.R
\name{prepare_agromet_API_data.fun}
\alias{prepare_agromet_API_data.fun}
\title{Prepare the data obtained by get_from_agromet_api.fun(). It types all the character variables to their proper types.}
\usage{
prepare_agromet_API_data.fun(meta_and_records.l, table_name.chr = NULL)
}
\arguments{
\item{meta_and_records.l}{a list containing agromet records and metadata returned by get_from_agromet_api.fun().}

\item{table_name.chr}{a character specifying the name of the agromet table from which the data where called using get_from_agromet_api.fun()}
}
\value{
a typed dataframe
}
\description{
Prepare the data obtained by get_from_agromet_api.fun(). It types all the character variables to their proper types.
}
