% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{source_files_recursively.fun}
\alias{source_files_recursively.fun}
\title{Recursively source all the function stored in a specific folder}
\usage{
source_files_recursively.fun(path.chr)
}
\arguments{
\item{path.chr}{a character specifying the path to the folder where to recursively look for R files to source}
}
\value{
sourced functions from the R files located in the specified folder
}
\description{
Inspired from https://stackoverflow.com/questions/32862426/load-all-files-from-folder-and-subfolders
}
\author{
Thomas Goossens - pokyah.github.io
}
